Add Quote to update-checkout-config.json
Recently, we've given up on https://github.com/apple/swift/pull/26708
which was trying to include Quote in the apple/swift:tensorflow build.

However, we still need a little bit of that PR in order to keep track
of which version of Quote the quoting logic in the compiler depends on.

It is somewhat awkward to clone Quote during the build and then not use
it afterwards, but that's the best that our current infrastructure
allows us.