[semantic-arc-opts] Change semantic arc opts to use a SmallBlotSetVector instead of a SmallSetVector.
Blotting is a trick with Map/Set vectors that involves one assigning None to the
vector at the index mapped to an entry instead of deleting the entry itself from
the vector. This ensures that we do not need to move elements of the vector down
when erasing elements from the worklist.