Allow @qualified types
When describing a type, you can prefix it with the @qualified attribute to indicate that it is already fully qualified—that is, the first identifier is the module it can be found in. This can allow you to address types and modules that have been shadowed by identically-named declarations.

This commit includes a test, but some parts of it do not pass yet.