Remove warnings from TestJSONEncoder compilation.
Since the introduction of #26357/SR-11160 the TestJSONEncoder seems to
have been emitting the new warnings related to tuples being incorrectly
pattern matched.

The changes remove the incorrect code (removing the parenthesis around
the pattern).

Should not have any other effect (and the effect was not normally
visible if the test was successful, anyway).