ClangImporter: repair build on Windows
Explicitly apply the std::move as the signature of the callee is:
  void swift::Evaluator::cacheOutput<swift::ExtendedNominalRequest, _>(const Request &, swift::NominalTypeDecl *&&)

The second formal parameter here is a `swift::ClassDecl` which is a
subclass of NominalTypeDecl, however, the cast will construct a
`swift::NominalTypeDecl *`.  This is a lvalue and cannot be bound to a
r-value reference.  Explicitly construct the r-value reference via
`std::move`.