[ownership] Now that we can run -Onone tests with both ownership and without, run pound_assert in both modes to prevent this test from breaking.
I also extracted out the recursion test into a FileCheck test so that we can
avoid having to do a -verify pattern match again the exact line where the
recursion limit was hit. This limit is noisy in the face of codegen changes
since we are IIRC counting instructions. Thus using -verify we can't pattern
match since the line can change easily.