[Serialization] Recover from conformances with missing decls
Components of a requirement may be hidden behind an implementation-only
import. Attemps at deserializing them would fail on a 'module not
loaded' error. We only see failures in non-compilation paths, either in
indexing or with tools like ide-test because they try to deserialize
things that are private.

Fix rdar://problem/60291019