[Diagnostics] Introduce the concept of a "public diagnostic".
A public diagnostic has a unique, stable identifier which is shown to
the user and is appropriate for looking up documentation, searching
Google/Stack Overflow, etc. A single public diagnostic corresponds to
one or more internal diagnostics. It also carries information about
the compiler version it was added/removed in. In the future, it may also
have its own associated documentation or reference material.