[NFC] OpaqueValueExpr::isPlaceholder() -> canBeCaptured()
This commit renames “isPlaceholder” to “canBeCaptured” and removes invariants stating that only “isPlaceholder” OVEs can appear in un-typechecked code. It turns out that we will want some “placeholder” OVEs to not be capturable, and the exact set of OVEs that should never appear in un-typechecked code is not a concept we need outside of assertions.