Remove the $interpolation variable
Previously, TapExpr used an actual, physical VarDecl to store the value it was modifying. That allowed us to represent references to that value as DeclRefExprs, which gave us some type-checking functionality for free but introduced issues like scoping that were not really necessary. This commit redesigns TapExpr to instead use OpaqueValueExprs as stand-ins for the value being modified. The backing storage is not directly represented in the AST; instead it is emitted in SILGen.

This commit changes two existing tests:

* DebugInfo/StringInterpolation.swift tested that the $interpolation variable was marked as artificial so it would be ignored. There is no such variable anymore, so it is now unnecessary.
* SILGen/sil_locations.swift had one SourceLoc move from inside the string literal to before it. This is fine.

IDE/complete_at_top_level.swift is broken in this commit, but will be fixed in a future commit.