[NFC] Simplify Opaque Type Registration
Drop the table of unvalidated declarations and rename the accessor so
it's a little clearer what kind of lookup this will perform.  On lookup,
the declaration will be validated via the call to getInterfaceType().

In the future, we should investigate making this even lazier by letting
the client control where validation should occur unless we're absolutely
certain this entrypoint should return a fully validated declaration.