Requestify Operator Validation
Restructure the way operators and precedencegroup declarations are validated.  First, factor out the common operator validation code into the DeclChecker.  Then, redo the lookup and validation code for precedence groups to work with the request model.  Finally, delete the validateDecl overloads on the TypeChecker.

Unfortunately, the evaluator is not capable of detecting all the cycles TypeCheckDecl can.  In particular, certain cross-file precedence group cycles cannot be diagnosed via  request alone.  That infrastructure all stays in place.