[lit] Silenced notes/warnings on quiet.
Lit can take a "quiet" option, -q, which is documented to "suppress no
error output".  Previously, test/lit.cfg displayed notes and warnings
when the quiet option was specified.  The result was that it was not
possible to get only pertinent file/line information to be used by an
editor to jump to the location where checks were failing without passing
a number of unhelpful locations first.  Here, all notes and warnings
displayed by test/lit.cfg are suppressed if the quiet option is set.