[stdlib] Remove UnsafeAtomicBool
It’s not layout-compatible with an actual Bool type, so this formulation is obviously broken. (D’oh.)

We could switch to single byte atomics, but it doesn’t seem worth the complexity. (Also, it’s unclear to me if that would have unusual alignment expectations or if it would add interference issues.)