[AST] Add Type::wrapInPointer(PointerTypeKind)
...replacing a similar utility in the type checker and one I just
added in ClangImporter earlier in this patch series. This is an
ever-so-slight regression in functionality because a missing pointer
declaration used to print a diagnostic and now it will just crash on
an invalid standard library, but I don't think we should be optimizing
for an invalid standard library anyway.