[Runtime] Remove bogus data from NOTES section
The "notes" section of an ELF binary is structured data and therefore
Swift must not place unrelated data into this section.

More importantly, if external programs want to introspect a binary, they
should (as the code does today), walk the ELF *sections* looking for the
data in question. Code must not walk ELF binaries looking for magic
data.